@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -26,10 +26,10 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.3
+  cupertino_icons: ^1.0.6
 
-  flutter_svg: ^0.22.0
-  get: ^4.3.8
+  flutter_svg: ^2.0.9
+  get: ^4.6.6
   # firebase_dynamic_links: ^2.0.9
   # url_launcher: ^6.0.12
 dev_dependencies:
