@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:news/controllers/MenuController.dart';
+import 'package:news/controllers/CustomMenuController.dart';
 import 'package:news/screens/home/home_screen.dart';
 
 import '../../constants.dart';
@@ -8,7 +8,8 @@ import 'components/header.dart';
 import 'components/side_menu.dart';
 
 class MainScreen extends StatelessWidget {
-  final MenuController _controller = Get.put(MenuController());
+  MainScreen({Key? key}) : super(key: key);
+  final CustomMenuController _controller = Get.put(CustomMenuController());
   @override
   Widget build(BuildContext context) {
     return Scaffold(
